*interesting commit message*

+++ task_2_raft/command_executor.py
@@ -1,24 +1,31 @@
 import asyncio
+import io
+import subprocess
 
 import common
 
 addr : str
 
-async def command_handler(reader: asyncio.StreamReader, writer: asyncio.StreamWriter):
+async def command_handler(reader: asyncio.StreamReader, writer: asyncio.StreamWriter) -> None:
     try:
-        command = reader.read()
-        process = await asyncio.create_subprocess_shell(command, stdout=writer, stderr=writer)
+        command = await reader.read()
+        if not command:
+            return
+        process = await asyncio.create_subprocess_shell(command, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
         try:
-            await process.communicate()
+            stdout, stderr = await process.communicate()
         finally:
             process.terminate()
-            await process.wait
+            await process.wait()
+        writer.write(stdout)
+        writer.write(stderr)
+        await writer.drain()
     finally:
-        common.safe_close_socket(writer)
+        await common.safe_socket_close(writer)
 
-async def main():
+async def main() -> None:
     async with await asyncio.start_server(command_handler, '0.0.0.0', 2100) as server:
-        server.serve_forever()
+        await server.serve_forever()
 
 if __name__ == '__main__':
     asyncio.run(main())
